refactor: standalone -> nodejs
