+++ docs/content/1.getting-started/3.examples.md
@@ -7,5 +7,5 @@ description: Example projects using tRPC Vue Query
 
 Here are some examples to help you get started with tRPC Vue Query:
 
-- [Standalone Example](https://github.com/falcondev-oss/trpc-vue-query/tree/master/examples/standalone) - A basic example showing how to use tRPC Vue Query in a standalone Vue.js application. [Click here to download the example](http://download-directory.github.io/?url=https%3A%2F%2Fgithub.com%2Ffalcondev-oss%2Ftrpc-vue-query%2Ftree%2Fmaster%2Fexamples%2Fstandalone)
+- [Standalone NodeJS Example](https://github.com/falcondev-oss/trpc-vue-query/tree/master/examples/nodejs) - A basic example showing how to use tRPC Vue Query in a standalone Vue.js application. [Click here to download the example](http://download-directory.github.io/?url=https%3A%2F%2Fgithub.com%2Ffalcondev-oss%2Ftrpc-vue-query%2Ftree%2Fmaster%2Fexamples%2Fnodejs)
 - [Nuxt Example](https://github.com/falcondev-oss/trpc-vue-query/tree/master/examples/nuxt) - An example showing how to use tRPC Vue Query in a Nuxt.js application. [Click here to download the example](http://download-directory.github.io/?url=https%3A%2F%2Fgithub.com%2Ffalcondev-oss%2Ftrpc-vue-query%2Ftree%2Fmaster%2Fexamples%2Fnuxt)
