+++ examples/nodejs/src/App.vue
@@ -1,0 +1,11 @@
+<script lang="ts" setup>
+import { useTRPC } from './useTRPC'
+
+const { data: greeting } = useTRPC().hello.useQuery({ input: 'world' })
+</script>
+
+<template>
+  <div>
+    <h1>{{ greeting }}</h1>
+  </div>
+</template>
