+++ examples/nodejs/src/vite-env.d.ts
