+++ examples/nodejs/tsconfig.json
@@ -1,0 +1,4 @@
+{
+  "references": [{ "path": "./tsconfig.app.json" }, { "path": "./tsconfig.node.json" }],
+  "files": []
+}
