--- examples/standalone/server/index.ts
@@ -1,25 +1,0 @@
-import { initTRPC } from '@trpc/server';
-import { createHTTPServer } from '@trpc/server/adapters/standalone';
-import { z } from 'zod';
-
-const t = initTRPC.create();
-
-const publicProcedure = t.procedure;
-const router = t.router;
-
-const appRouter = router({
-  hello: publicProcedure
-    .input(z.object({ input: z.string() }))
-    .query(({ input }) => {
-      return `Hello ${input.input}!`;
-    }),
-});
-
-export type AppRouter = typeof appRouter;
-
-const server = createHTTPServer({
-  router: appRouter,
-});
-
-server.listen(3000);
-console.log('Server running on http://localhost:3000');
